// src/interfaces/IData.tsx
export default interface IData {
  target: any;
  Numero: string
  Estado: string
  KilosFolio: string
  Cprod: string
  Grp_prod: string
  Chofer: string
  PesoCamion: string
  TaraCamion: string
  PatenteCarro: string
  TotEnvSal: string
  PatenteCamion: string
  FecRecepcion: string
  Transportista: string
  GuiaProductor: string
  FecCosecha: string
  TipoFruta: string
  Folio: string
  CHto: string
  SDP: string
  TranspRut: string
  Car1: string
  CEnv: string
  CTfr: string
  CCnd: string
  PesoCarro: string
  TaraCarro: string
  ConEnvases: string
  Bins: string
  Kilos_prom: string
  FecDestara: string
  Cesp: string
  Cvar: string
  Grp_Var: string
  CPlt: string
  Marca: string
  Nomb: string
  FecSalida: string
  FecPeso: string
  Tipo_cos: string
  UM: string
  Nom_prod: string
  Color: string
  Nota: string
  Temporada: string
  CSG: string
  }
  // RECORDAR, que si se usa en SQL la fecha, esta tiene un formato especifico. De no ser necesario, se puede cambiar a texto para evitar incompatibilidades.
  //NO CAMBIAR  LA FECHA A FORMATO TEXTO, YA QUE SE NECESITA PARA EL FILTRO DE FECHAS EN EL FRONTEND